pyproject.toml
[project]
name = "robust-llm"
description = "Testing the robustness of LLMs"
requires-python = ">=3.10"
version = "0.0.1"

dependencies = [
    "evaluate",
    "datasets",
    "huggingface_hub",
    "hydra-core",
    "torch",
    "transformers[accelerate]",
    "simple_parsing",
    "scikit-learn",
    # Bigger than this will break TextAttack
    "scipy<1.13.0",
    "tdigest",
    "textattack",
    "trl",
    "wandb",
    "semver",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pre-commit",
    "mypy",
    "pyright",
    # Needed for mypy to type check PyYAML
    "types-PyYAML",
    "pytest",
    "hypothesis",
    "pytest-cov",
    "pytest-mock",
    "pytest-xdist",
    "flake8",
    "names-generator",
    # For checking GPU memory usage
    "nvidia-ml-py3",
]

# Our models are all in PyTorch, but TextAttack uses TF embeddings for some attacks.
# Install TensorFlow if you want to run those attacks.
tensorflow = [
    "tensorflow>=2.9.1",
    "tensorflow_hub",
]

[tool.setuptools.packages.find]
namespaces = false

[tool.isort]
profile = "black"
# Necessary because of directories with
# packages with the same name
known_third_party = ["datasets", "trl", "wandb"]

[tool.pyright]
venvPath = "."
venv = "venv"
include = ["robust_llm"]
pythonVersion = "3.10"

[tool.mypy]
ignore_missing_imports = true
python_version = "3.10"
